Chart Display: Move demo to test directory.
